requirements/_base.txt
alabaster==0.7.3
amqp==1.4.6
anyjson==0.3.3
Babel==1.3
billiard==3.3.0.20
braintree==3.13.0
celery==3.1.17
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.8.0
django-celery==3.1.16
django-filter==0.9.2
django-redis==3.8.3
django-suit==0.2.13
djangorestframework==3.1.1
docutils==0.12
gunicorn==19.3.0
Jinja2==2.7.3
kombu==3.0.25
Markdown==2.6.2
MarkupSafe==0.23
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.12
Pygments==2.0.2
pystache==0.5.4
pytz==2015.2
redis==2.10.3
requests==2.6.0
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
static==1.1.1
static3==0.5.1
testfixtures==4.1.2

# ie: from dirtyfields import DirtyFieldsMixin
# models use this mixin to see what fields of an instance have changed
django-dirtyfields==0.7

# De-BREACH security (see http://goo.gl/KPTym3 for info)
django-debreach==1.1.0

# Django Extensions (shell plus, url map, dataviz of db, etc)
django-extensions==1.5.5

# Enhanced shell capabilities
ipython==2.3.1

# Sentry dependency
raven==5.1.1

# Paths
Unipath==1.1

# Compressing, serving static assets
django-pipeline==1.4.2
jsmin==2.0.11

# Static asset serving via wsgi
whitenoise==1.0.6

# Model caching
django-cachalot==1.0.0

# python api for mongodb
pymongo==3.0.1

# python-dateutil is a better datetime parser
python-dateutil==2.4.2

# Flower (celery task metrics/logging/captures)
flower==0.8.3
